pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "s3_serverless_inline_storage_utils"
version = "1.1.8"
description = "S3 Storage Utilities that do not rely on environments to interact with S3 instances. Work with your buckets directly from within the workflow."
authors = [
  {name = "Uday Jain", email = "uj9.uday@gmail.com"}
]
readme = "README.md"
license = {text = "GNU General Public License v3"}
repository = "https://github.com/udayjain26/s3_serverless_inline_storage_utils"
classifiers = []
dependencies = [
    "boto3>=1.26.0",
    "botocore>=1.29.0",
    "pillow>=10.0.0",
    "requests>=2.28.0",
]



[project.optional-dependencies]
dev = [
    "bump-my-version",
    "coverage",  # testing
    "mypy",  # linting
    "pre-commit", # runs linting on commit
    "pytest",  # testing
    "ruff",  # linting
]

[project.urls]
bugs = "https://github.com/udayjain26/s3_serverless_inline_storage_utils/issues"
homepage = "https://github.com/udayjain26/s3_serverless_inline_storage_utils"
Repository = "https://github.com/udayjain26/s3_serverless_inline_storage_utils"


[tool.comfy]
PublisherId = "ascension"
DisplayName = "S3 Serverless-Inline Storage Utils"
Icon = ""




[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[tool.ruff]
# extend-exclude = ["static", "ci/templates"]
line-length = 140
src = ["src", "tests"]
target-version = "py39"

# Add rules to ban exec/eval
[tool.ruff.lint]
select = [
    "S102",  # exec-builtin
    "S307",  # eval-used
    "W293",
    "F",  # The "F" series in Ruff stands for "Pyflakes" rules, which catch various Python syntax errors and undefined names.
    # See all rules here: https://docs.astral.sh/ruff/rules/#pyflakes-f
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
